Refactor function
